tests:Move more button functions
