+++ TestApp/lib/TestFunctions.cs
@@ -1124,50 +1124,5 @@ public static class TestFunctions
 
 
 
-    private class Model_ClickCountInButtonWithTypedDataContext : nac.Forms.model.ViewModelBase
-    {
-        public int Count
-        {
-            get { return GetValue(() => Count); }
-            set { SetValue(() => Count, value);}
-        }
-
-        public int Count2
-        {
-            get { return GetValue(() => Count2); }
-            set { SetValue(() => Count2, value);}
-        }
-    }
-    public static void TestButton_ClickCountInButtonWithTypedDataContext(Form f)
-    {
-        var model = new Model_ClickCountInButtonWithTypedDataContext();
-        f.DataContext = model;
-
-        f.HorizontalGroup(buttonRow =>
-            buttonRow.Button(_b => _b
-                        .HorizontalGroup(hg =>
-                        {
-                            hg.Text("Click (")
-                                .TextFor(nameof(model.Count),
-                                    style: new nac.Forms.model.Style
-                                        { foregroundColor = Avalonia.Media.Colors.Red })
-                                .Text(")");
-                        }),
-                    async () =>
-                    {
-                        ++model.Count;
-                    }
-                )
-                .Button(_b => _b.HorizontalGroup(hg =>
-                {
-                    hg.Text(" Click-2 (")
-                        .TextFor(nameof(model.Count2),
-                            style: new Style { foregroundColor = Avalonia.Media.Colors.Green })
-                        .Text(")");
-                }), async () =>
-                {
-                    ++model.Count2;
-                })
-        ); // end of horizontal group button row
-    }
+    
 }
